Extract main module for activity
