+++ main/src/main/java/com/bhaktaprogram/notehelper/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.bhaktaprogram.notehelper
+package com.bhaktaprogram.notehelper.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
